extract common prompt context logic
